+++ tests/common/for_each_gaussian_pod.rs
@@ -4,17 +4,17 @@ macro_rules! for_each_gaussian_pod {
         fn _body<$pod: wgpu_3dgs_core::GaussianPod>() {
             $body
         }
-        _body::<wgpu_3dgs_core::GaussianPodWithShSingleCov3dRotScaleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShSingleCov3dSingleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShSingleCov3dHalfConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShHalfCov3dRotScaleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShHalfCov3dSingleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShHalfCov3dHalfConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShNorm8Cov3dRotScaleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShNorm8Cov3dSingleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShNorm8Cov3dHalfConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShNoneCov3dRotScaleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShNoneCov3dSingleConfigs>();
-        _body::<wgpu_3dgs_core::GaussianPodWithShNoneCov3dHalfConfigs>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShSingle, wgpu_3dgs_core::CovRotScale>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShSingle, wgpu_3dgs_core::CovSingle>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShSingle, wgpu_3dgs_core::CovHalf>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShHalf, wgpu_3dgs_core::CovRotScale>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShHalf, wgpu_3dgs_core::CovSingle>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShHalf, wgpu_3dgs_core::CovHalf>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShNorm8, wgpu_3dgs_core::CovRotScale>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShNorm8, wgpu_3dgs_core::CovSingle>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShNorm8, wgpu_3dgs_core::CovHalf>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShNone, wgpu_3dgs_core::CovRotScale>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShNone, wgpu_3dgs_core::CovSingle>>();
+        _body::<wgpu_3dgs_core::PackedGaussian<wgpu_3dgs_core::ShNone, wgpu_3dgs_core::CovHalf>>();
     };
 }